moved resource hand and methods correlated
