+++ client/src/main/resources/fxml/playersBoard.fxml
@@ -12,6 +12,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
+<?import javafx.scene.control.Pagination?>
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="668.0" prefWidth="936.0" styleClass="anchor" stylesheets="@../css/playersBoard.css" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="it.polimi.ingsw.client.GUI.BoardController">
     <children>
 
@@ -155,5 +156,7 @@
             <ImageView fx:id="activeLeaderCard2" fitHeight="260.0" fitWidth="200.0" layoutX="25.0" layoutY="441.0" pickOnBounds="true" preserveRatio="true" />
             <Label fx:id="labelLeaderCards" layoutX="56.0" layoutY="119.0" prefHeight="27.0" prefWidth="139.0" text="Leader Cards:" />
         </Pane>
+        <!--resource hand -->
+        <Pagination fx:id="hand" layoutX="22.0" layoutY="139.0" maxPageIndicatorCount="0" onDragDetected="#moveRes" pageCount="0" prefHeight="151.0" prefWidth="204.0" />
     </children>
 </AnchorPane>
